requirements.txt
alpha-vantage==2.0.0
numpy>=1.16.2
pandas>=0.23.0
decorator>=4.1.2
tqdm>=4.32.1
matplotlib>=2.1.0
statsmodel>=0.10.1
scipy>=1.3.0
patsy>=0.5.1
datetools>=1.1
sklearn>=0.21.2
seaborn>=0.9.0
